Containers for each check subcommand
