+++ cmd/actuary/server/results/results.html
@@ -3,6 +3,8 @@
 <head>
 	<title>Actuary</title>
 	<meta charset="utf-8" name="viewport" content="width=device-width, initial-scale=1">
+	<script type="text/javascript" src = "js/print_results.js"> </script>
+	<script type="text/javascript" src = "js/analyze_results.js"> </script>
 	<script type="text/javascript" src = "js/main.js"> </script>
 	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
 	<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
