+++ docker-compose.yml
@@ -28,7 +28,7 @@ services:
       X509_USER_KEY: /run/secrets/tls_key
       X509_USER_CERT: /run/secrets/tls_cert
       TOKEN_PASSWORD: /run/secrets/token_password
-    command: ["check", "-f=cmd/actuary/mac-default.toml", "-s=https://server:8000/results"]
+    command: ["check", "-f=cmd/actuary/mac-default.toml", "-p=/go/src/github.com/diogomonica/actuary", "-s=https://server:8000/results", "--swarm=true"]
     deploy:
       mode: global
       resources:
@@ -55,3 +55,4 @@ secrets:
     file: ./domain.key
   token_password:
     file: ./token_password.txt
+
