+++ vendor.conf.tmp
@@ -29,4 +29,4 @@ github.com/spf13/cobra c46add8a652801b61513ad36c56759f302fbb028
 github.com/spf13/pflag e57e3eeb33f795204c1ca35f56c44f83227c6e66
 github.com/dgrijalva/jwt-go a539ee1a749a2b895533f979515ac7e6e0f5b650
 github.com/gorilla/context 08b5f424b9271eedf6f9f0ce86cb9396ed337a42
-
+github.com/moby/moby 33d82b78d03f7db4ef743951c6a004ba4fe64637
